+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.1
 
+* Schema objects can now be resolved from classpath using `DDLUnit.fromClasspath()` 
+
 * Migrated to Scala 2.9
 
 * Method naming (introduce incompatibilities!):